+++ src/main/java/com/telepathicgrunt/blame/mixin/JigsawManagerMixin.java
@@ -1,5 +1,6 @@
 package com.telepathicgrunt.blame.mixin;
 
+import com.telepathicgrunt.blame.main.JigsawStartPoolBlame;
 import com.telepathicgrunt.blame.main.MissingNBTBlame;
 import net.minecraft.util.Rotation;
 import net.minecraft.util.math.BlockPos;
@@ -8,6 +9,7 @@ import net.minecraft.util.registry.MutableRegistry;
 import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.feature.jigsaw.JigsawManager;
 import net.minecraft.world.gen.feature.jigsaw.JigsawPattern;
+import net.minecraft.world.gen.feature.jigsaw.JigsawPiece;
 import net.minecraft.world.gen.feature.structure.AbstractVillagePiece;
 import net.minecraft.world.gen.feature.structure.VillageConfig;
 import net.minecraft.world.gen.feature.template.TemplateManager;
@@ -30,14 +32,15 @@ import java.util.Random;
 public class JigsawManagerMixin {
 
     @Inject(method = "addPieces(Lnet/minecraft/util/registry/DynamicRegistries;Lnet/minecraft/world/gen/feature/structure/VillageConfig;Lnet/minecraft/world/gen/feature/jigsaw/JigsawManager$IPieceFactory;Lnet/minecraft/world/gen/ChunkGenerator;Lnet/minecraft/world/gen/feature/template/TemplateManager;Lnet/minecraft/util/math/BlockPos;Ljava/util/List;Ljava/util/Random;ZZ)V",
-            at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/gen/feature/jigsaw/JigsawPattern;getRandomTemplate(Ljava/util/Random;)Lnet/minecraft/world/gen/feature/jigsaw/JigsawPiece;"),
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/gen/feature/jigsaw/JigsawPattern;getRandomTemplate(Ljava/util/Random;)Lnet/minecraft/world/gen/feature/jigsaw/JigsawPiece;"),
             locals = LocalCapture.CAPTURE_FAILHARD)
     private static void storeCurrentPool(DynamicRegistries dynamicRegistries, VillageConfig villageConfig,
                                          JigsawManager.IPieceFactory iPieceFactory, ChunkGenerator chunkGenerator,
                                          TemplateManager templateManager, BlockPos blockPos,
-                                         List<? super AbstractVillagePiece> pieces_list, Random random,
+                                         List<? super AbstractVillagePiece> piecesList, Random random,
                                          boolean b, boolean b1, CallbackInfo ci, MutableRegistry<JigsawPattern> mutableregistry,
                                          Rotation rotation, JigsawPattern jigsawpattern) {
+        JigsawStartPoolBlame.printStartPoolCrash(jigsawpattern);
         MissingNBTBlame.CALLING_POOL = jigsawpattern.getName();
     }
 }